+++ src/factories/event.ts
@@ -1,5 +1,6 @@
 import {
   Actor,
+  AdditionalPhonemeInfo as ProtoAdditionalPhonemeInfo,
   CancelResponsesEvent,
   ControlEvent,
   CustomEvent,
@@ -9,6 +10,7 @@ import {
   Routing,
   TextEvent,
 } from '@proto/packets_pb';
+import { Duration } from 'google-protobuf/google/protobuf/duration_pb';
 import { v4 } from 'uuid';
 
 import { protoTimestamp } from '../common/helpers';
@@ -17,6 +19,7 @@ import { Character } from '../entities/character.entity';
 import { EmotionBehavior } from '../entities/emotion-behavior.entity';
 import { EmotionStrength } from '../entities/emotion-strength.entity';
 import {
+  AdditionalPhonemeInfo,
   InworlControlType,
   InworldPacket,
   InworldPacketType,
@@ -87,19 +90,21 @@ export class EventFactory {
     return this.protoPacket().setCancelresponses(event);
   }
 
-  convertToInworldPacket(packet: ProtoPacket): InworldPacket {
-    const packetId = packet.getPacketId();
-    const routing = packet.getRouting();
+  static fromProto(proto: ProtoPacket): InworldPacket {
+    const packetId = proto.getPacketId();
+    const routing = proto.getRouting();
     const source = routing.getSource();
     const target = routing.getTarget();
-    const type = this.getType(packet);
+    const type = this.getType(proto);
 
-    const textEvent = packet.getText();
-    const emotionEvent = packet.getEmotion();
+    const textEvent = proto.getText();
+    const emotionEvent = proto.getEmotion();
+    const additionalPhonemeInfo =
+      proto.getDataChunk()?.getAdditionalPhonemeInfoList() ?? [];
 
     return new InworldPacket({
       type,
-      date: packet.getTimestamp().toDate().toISOString(),
+      date: proto.getTimestamp().toDate().toISOString(),
       packetId: {
         packetId: packetId.getPacketId(),
         utteranceId: packetId.getUtteranceId(),
@@ -119,7 +124,7 @@ export class EventFactory {
       },
       ...(type === InworldPacketType.TRIGGER && {
         trigger: {
-          name: packet.getCustom().getName(),
+          name: proto.getCustom().getName(),
         },
       }),
       ...(type === InworldPacketType.TEXT && {
@@ -130,12 +135,24 @@ export class EventFactory {
       }),
       ...(type === InworldPacketType.AUDIO && {
         audio: {
-          chunk: packet.getDataChunk().getChunk_asB64(),
+          chunk: proto.getDataChunk().getChunk_asB64(),
+          additionalPhonemeInfo: additionalPhonemeInfo.map(
+            (info: ProtoAdditionalPhonemeInfo) =>
+              ({
+                phoneme: info.getPhoneme(),
+                startOffsetS: this.durationToSeconds(info.getStartOffset()),
+              } as AdditionalPhonemeInfo),
+          ),
         },
       }),
       ...(type === InworldPacketType.CONTROL && {
         control: {
-          type: this.getControlType(packet),
+          type: this.getControlType(proto),
+        },
+      }),
+      ...(type === InworldPacketType.SILENCE && {
+        silence: {
+          durationMs: proto.getDataChunk().getDurationMs(),
         },
       }),
       ...(type === InworldPacketType.EMOTION && {
@@ -146,8 +163,8 @@ export class EventFactory {
       }),
       ...(type === InworldPacketType.CANCEL_RESPONSE && {
         cancelResponses: {
-          interactionId: packet.getCancelresponses().getInteractionId(),
-          utteranceId: packet.getCancelresponses().getUtteranceIdList(),
+          interactionId: proto.getCancelresponses().getInteractionId(),
+          utteranceId: proto.getCancelresponses().getUtteranceIdList(),
         },
       }),
     });
@@ -165,7 +182,7 @@ export class EventFactory {
 
     const target = new Actor()
       .setType(Actor.Type.AGENT)
-      .setName(this.character?.getId());
+      .setName(this.character?.id);
 
     return new Routing().setSource(source).setTarget(target);
   }
@@ -174,13 +191,16 @@ export class EventFactory {
     return new PacketId().setPacketId(v4());
   }
 
-  private getType(packet: ProtoPacket) {
+  private static getType(packet: ProtoPacket) {
     switch (true) {
       case packet.hasText():
         return InworldPacketType.TEXT;
       case packet.hasDataChunk() &&
         packet.getDataChunk().getType() === DataChunk.DataType.AUDIO:
         return InworldPacketType.AUDIO;
+      case packet.hasDataChunk() &&
+        packet.getDataChunk().getType() === DataChunk.DataType.SILENCE:
+        return InworldPacketType.SILENCE;
       case packet.hasCustom():
         return InworldPacketType.TRIGGER;
       case packet.hasControl():
@@ -194,12 +214,16 @@ export class EventFactory {
     }
   }
 
-  private getControlType(packet: ProtoPacket) {
+  private static getControlType(packet: ProtoPacket) {
     switch (packet.getControl().getAction()) {
       case ControlEvent.Action.INTERACTION_END:
         return InworlControlType.INTERACTION_END;
       default:
         return InworlControlType.UNKNOWN;
     }
+  }
+
+  private static durationToSeconds(duration: Duration) {
+    return duration.getSeconds() + duration.getNanos() / 1000000000;
   }
 }