+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2022-04-20 v1.2.0
+
+* Add phonemes event support
+* Add silence event support
+* Improve manual reconnect latency
+
 ## 2022-04-06 v1.1.8
 
 * Remove deprecated emotions attributes: joy, fear, trust and surprise